requirements.txt
sanic==22.6.2
transformers
accelerate
git+https://github.com/openai/whisper.git 
--find-links https://download.pytorch.org/whl/torch_stable.html

torchaudio==0.12.1